changes chat utils unit tests, removes descirbe block and renames to pushMessageToHistory
